requirements.txt
discord
python-dotenv
motor
dnspython